+++ includes/security/Token.php
@@ -37,8 +37,8 @@ class Token extends Processor\ProcessorBase {
   );
 
   /**
-   * @return bool
-   * @throws \Datagator\Security\ApiException
+   * @return array
+   * @throws \Datagator\Core\ApiException
    */
   public function process() {
     Core\Debug::variable($this->meta, 'Validator TokenConsumer', 4);
@@ -48,7 +48,7 @@ class Token extends Processor\ProcessorBase {
 
     $user = $userMapper->findBytoken($token);
     if (empty($user->getUid()) || $user->getActive() == 0) {
-      throw new ApiException('permission denied', 4, -1, 401);
+      throw new Core\ApiException('permission denied', 4, -1, 401);
     }
 
     return array('result' => true);